backend/requirements.txt
aiohttp
aiohttp_cors
python-socketio
aiortc
ultralytics
opencv-python-headless
numpy
shapely
shapely
